# lib/api-zod/src/generated/types/roleInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Staff Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface RoleInput {
  /**
     * @minLength 1
     * @maxLength 50
     */
  name: string;
}